+++ packages/stargate/src/modules/authz/queries.ts
@@ -1,4 +1,9 @@
-import { QueryClientImpl, QueryGrantsResponse } from "cosmjs-types/cosmos/authz/v1beta1/query";
+import {
+  QueryClientImpl,
+  QueryGranteeGrantsResponse,
+  QueryGranterGrantsResponse,
+  QueryGrantsResponse,
+} from "cosmjs-types/cosmos/authz/v1beta1/query";
 
 import { createPagination, createProtobufRpcClient, QueryClient } from "../../queryclient";
 
@@ -10,6 +15,14 @@ export interface AuthzExtension {
       msgTypeUrl: string,
       paginationKey?: Uint8Array,
     ) => Promise<QueryGrantsResponse>;
+    readonly granteeGrants: (
+      grantee: string,
+      paginationKey?: Uint8Array,
+    ) => Promise<QueryGranteeGrantsResponse>;
+    readonly granterGrants: (
+      granter: string,
+      paginationKey?: Uint8Array,
+    ) => Promise<QueryGranterGrantsResponse>;
   };
 }
 
@@ -22,14 +35,24 @@ export function setupAuthzExtension(base: QueryClient): AuthzExtension {
   return {
     authz: {
       grants: async (granter: string, grantee: string, msgTypeUrl: string, paginationKey?: Uint8Array) => {
-        const response = await queryService.Grants({
+        return await queryService.Grants({
           granter: granter,
           grantee: grantee,
           msgTypeUrl: msgTypeUrl,
           pagination: createPagination(paginationKey),
         });
-
-        return response;
+      },
+      granteeGrants: async (grantee: string, paginationKey?: Uint8Array) => {
+        return await queryService.GranteeGrants({
+          grantee: grantee,
+          pagination: createPagination(paginationKey),
+        });
+      },
+      granterGrants: async (granter: string, paginationKey?: Uint8Array) => {
+        return await queryService.GranterGrants({
+          granter: granter,
+          pagination: createPagination(paginationKey),
+        });
       },
     },
   };